requirements.txt
boto3==1.28.1
factory-boy==3.2.0
Faker==13.15.1
fastapi==0.95.2
httptools==0.5.0
httpx==0.24.1
line-bot-sdk==3.1.0
orjson==3.9.9
pydantic-factories==1.4.1
PyMySQL==1.0.3
pytest==7.3.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.25
typer==0.7.0
uvicorn==0.22.0
watchtower==3.0.1
